ind_det/requirements.txt
tensorflow-cpu==2.5.0
numpy==1.19.5
Pillow==8.2.0
scikit_learn==0.24.2